Fixing exception in MAUI app
