+++ src/FairPlaySocialSln/FairPlaySocial.Common/Handlers/LocalizationMessageHandler.cs
@@ -1,0 +1,28 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Net.Http.Headers;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace FairPlaySocial.Common.Handlers
+{
+    public class LocalizationMessageHandler : DelegatingHandler
+    {
+        protected override HttpResponseMessage Send(HttpRequestMessage request, CancellationToken cancellationToken)
+        {
+            var currentCulture = System.Globalization.CultureInfo.CurrentUICulture;
+            request.Headers.AcceptLanguage.Clear();
+            request.Headers.AcceptLanguage.Add(new StringWithQualityHeaderValue(currentCulture.Name));
+            return base.Send(request, cancellationToken);
+        }
+
+        protected override async Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
+        {
+            var currentCulture = System.Globalization.CultureInfo.CurrentUICulture;
+            request.Headers.AcceptLanguage.Clear();
+            request.Headers.AcceptLanguage.Add(new StringWithQualityHeaderValue(currentCulture.Name));
+            return await base.SendAsync(request, cancellationToken);
+        }
+    }
+}
