+++ src/FairPlaySocialSln/FairPlaySocial/Client/Program.cs
@@ -5,6 +5,7 @@ using FairPlaySocial.Client.Extensions;
 using FairPlaySocial.Client.Services;
 using FairPlaySocial.ClientsConfiguration;
 using FairPlaySocial.ClientServices.CustomLocalization.Api;
+using FairPlaySocial.Common.Handlers;
 using FairPlaySocial.Common.Interfaces.Services;
 using Microsoft.AspNetCore.Components.Web;
 using Microsoft.AspNetCore.Components.WebAssembly.Authentication;
@@ -75,22 +76,3 @@ builder.Services.AddMsalAuthentication<RemoteAuthenticationState, CustomRemoteUs
 var host = builder.Build();
 await host.SetDefaultCulture();
 await host.RunAsync();
-
-public class LocalizationMessageHandler : DelegatingHandler
-{
-    protected override HttpResponseMessage Send(HttpRequestMessage request, CancellationToken cancellationToken)
-    {
-        var currentCulture = System.Globalization.CultureInfo.CurrentUICulture;
-        request.Headers.AcceptLanguage.Clear();
-        request.Headers.AcceptLanguage.Add(new StringWithQualityHeaderValue(currentCulture.Name));
-        return base.Send(request, cancellationToken);
-    }
-
-    protected override async Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
-    {
-        var currentCulture = System.Globalization.CultureInfo.CurrentUICulture;
-        request.Headers.AcceptLanguage.Clear();
-        request.Headers.AcceptLanguage.Add(new StringWithQualityHeaderValue(currentCulture.Name));
-        return await base.SendAsync(request, cancellationToken);
-    }
-}
